fix GlobalChoreProvider world handling
The fetchMap has parentWorldId as the indexes, and since parent is set to point
to self if a world has no parent, using merely parentWorldId is enough, as its
the same id for a world and its subworlds.

This mostly reverts afee736ebc114312b4d34c447bbe3342703833b5 and fixes it
in a simpler way.